temp

+++ OpenNefia.Content/CharaMake/CharaMakeLayer.cs
@@ -102,13 +102,13 @@ namespace OpenNefia.Content.CharaMake
             Data = args;
         }
 
-        public override void SetSize(int width, int height)
+        public override void SetSize(float width, float height)
         {
             base.SetSize(width, height);
             Caption.SetPreferredSize();
         }
 
-        public override void SetPosition(int x, int y)
+        public override void SetPosition(float x, float y)
         {
             base.SetPosition(x, y);
             Caption.SetPosition(20, 30);
