+++ OpenNefia.Content/ConfigMenu/UICell/BaseConfigMenuUICell.cs
@@ -103,7 +103,7 @@ namespace OpenNefia.Content.ConfigMenu.UICell
             AssetArrowRight.Color = rightArrowEnabled ? Color.White : ColorArrowDisabled;
         }
 
-        public override void SetSize(int width, int height)
+        public override void SetSize(float width, float height)
         {
             base.SetSize(width, height);
             ValueText.SetPreferredSize();
@@ -111,7 +111,7 @@ namespace OpenNefia.Content.ConfigMenu.UICell
             AssetArrowRight.SetPreferredSize();
         }
 
-        public override void SetPosition(int x, int y)
+        public override void SetPosition(float x, float y)
         {
             base.SetPosition(x, y);
             ValueText.SetPosition(X + 194, Y + 1);
