+++ OpenNefia.Content/UI/Element/UiWindow.cs
@@ -84,7 +84,7 @@ namespace OpenNefia.Content.UI.Element
             TextKeyHint.Text = UserInterfaceManager.FormatKeyHints(KeyHints);
         }
 
-        public override void SetPosition(int x, int y)
+        public override void SetPosition(float x, float y)
         {
             base.SetPosition(x, y);
 
@@ -103,7 +103,7 @@ namespace OpenNefia.Content.UI.Element
             TextKeyHint.SetPosition(x + 58 + KeyHintXOffset, y + Height - 43 - Height % 8);
         }
 
-        public override void SetSize(int width, int height)
+        public override void SetSize(float width, float height)
         {
             base.SetSize(width, height);
